Convert eval command to kotlin-engine and fix dashboard

+++ shared/src/main/java/com/jagrosh/jdautilities/oauth2/entities/OAuth2Guild.java
@@ -66,7 +66,7 @@ public interface OAuth2Guild extends ISnowflake
      *
      * @return The Session User's raw permission value for the Guild.
      */
-    int getPermissionsRaw();
+    long getPermissionsRaw();
 
     /**
      * Gets the Session User's {@link Permission Permissions} for the Guild.
